doc: add git-psuh help page
This new document enables the use of `git help psuh`, used by users
when trying out a new command for the first time.

The name of the help page (git-FOO.txt) ensures it is picked up when
documentation is being built. It will be processed by AsciiDoc as well
as AsciiDoctor; each tool should produce nice results.

Signed-off-by: Carlos Pavanetti <carlospavanetti.silvaprado@gmail.com>